support background clipboard waiting, add clap CLI args
Apparently something in Gnome immediately grabs things from the
clipboard so we don't have to wait, but a bare X11 server like with WSL2
(and possibly SSH forwarding, though I haven't tested that) we do
actually need to stay alive to service clipboard requests.

arboard's "daemonize" example re-execs itself with a special argument,
but fw is simple enough that we can just use a single fork() safely.

Also bring in clap and add some options to control whether and how we
save to the clipboard.